==> notebooks/2-share-and-offer-calculations.ipynb <==
import pandas as pd
import numpy as np

info = pd.read_csv('../data/cleaned/info.csv')
info = info.set_index('dbn')
directory = pd.read_csv('../data/2020_DOE_High_School_Directory.csv')

metrics = pd.read_csv('../data/metrics_2019.csv')
metrics = metrics.set_index('dbn')

school_names = pd.read_csv('../data/cleaned/school_names.csv')
school_names = school_names.set_index('dbn')

folder = '../data/cleaned/'

applicants = pd.read_csv(folder + 'applicants.csv')
applicants = applicants.set_index('dbn')

avg_applicants = pd.read_csv(folder + 'avg_applicants.csv')
avg_applicants = avg_applicants.set_index('dbn')

min_applicants = pd.read_csv(folder + 'min_applicants.csv')
min_applicants = min_applicants.set_index('dbn')

max_applicants = pd.read_csv(folder + 'max_applicants.csv')
max_applicants = max_applicants.set_index('dbn')

offers = pd.read_csv(folder + 'offers.csv')
offers = offers.set_index('dbn')

avg_offers = pd.read_csv(folder + 'avg_offers.csv')
avg_offers = avg_offers.set_index('dbn')

min_offers = pd.read_csv(folder + 'min_offers.csv')
min_offers = min_offers.set_index('dbn')

max_offers = pd.read_csv(folder + 'max_offers.csv')
max_offers = max_offers.set_index('dbn')

cols = ['Total', 'Asian','Black','Hispanic','White']

### Offer Rates

#### Individual Schools

avg_offers.div(avg_applicants)[cols]

#### All Schools

avg_offers.sum().div(avg_applicants.sum())[cols]

#### Top Ranking Schools

top = metrics[metrics.eq('Far Above Average').sum(axis=1).gt(4)].index
print(len(top))
top = [dbn for dbn in top if dbn in info.index]
print(len(top))

avg_offers.loc[top].sum().div(avg_applicants.loc[top].sum())[cols]

df_info = pd.DataFrame([['All Screened Schools', len(offers), applicants['Total'].sum(), offers['Total'].sum()],
              ['"Top-Ranked" Screened Schools', len(offers.loc[top]), applicants.loc[top]['Total'].sum(), offers.loc[top]['Total'].sum()]], columns=['index','No. of Schools', 'No. of Applicants', 'No. of Offers'])
df_info = df_info.set_index('index')

df = pd.DataFrame()
df['All Screened Schools'] = avg_offers.sum().div(avg_applicants.sum()).mul(100).round(1)[cols]
df['"Top-Ranked" Screened Schools'] = avg_offers.loc[top].sum().div(avg_applicants.loc[top].sum()).mul(100).round(1)[cols]
df = df.T
df.columns = [col.replace('Hispanic', 'Latino').replace('Total','Overall') + " Offer Rate (%)" for col in df.columns]

display(df)

df_info.join(df, how='outer').to_csv('../output/all-vs-top-ranked-rates.csv')
df_info.join(df, how='outer')

for col in cols:
    count = avg_offers.loc[top].sum()[col]
    nobs  = avg_applicants.loc[top].sum()[col]
    prop_var = min_offers.loc[top].sum()[col] / max_applicants.loc[top].sum()[col]
    value = 0.05
    stats, pval = proportions_ztest(count, nobs, value, prop_var = prop_var)
    print(col, '{0:0.3f}'.format(pval))

    prop_var = max_offers.loc[top].sum()[col] / min_applicants.loc[top].sum()[col]
    stats, pval = proportions_ztest(count, nobs, value, prop_var = prop_var)
    print(col, '{0:0.3f}'.format(pval))
    
    print(round(avg_offers.loc[top].sum()[col] / avg_applicants.loc[top].sum()[col] * 100, 1),
          round(min_offers.loc[top].sum()[col] / max_applicants.loc[top].sum()[col] * 100, 1),
          round(max_offers.loc[top].sum()[col] / min_applicants.loc[top].sum()[col] * 100, 1))

### Demographic Breakdowns of Applicants and Offers

#### Individual Schools

redacted_offers = offers.eq('s') | offers.eq('s^')
redacted_applicants = applicants.eq('s') | applicants.eq('s^')

df = pd.DataFrame()
df['School Name'] = school_names['school_name']
df['# Applicants'] = avg_applicants['Total']
df[['% Asian Applicants', '% Black Applicants', '% Latino Applicants', '% White Applicants']] = avg_applicants[['Asian', 'Black', 'Hispanic', 'White']].div(avg_applicants['Total'],axis='index').mul(100).round(1)
df[['% Asian Applicants - Estimate Off By (+/-)', '% Black Applicants - Estimate Off By (+/-)', '% Latino Applicants - Estimate Off By (+/-)', '% White Applicants - Estimate Off By (+/-)']] = max_applicants.sub(avg_applicants).div(avg_applicants['Total'], axis='index').mul(100).round(1)[['Asian', 'Black','Hispanic','White']]
df[['# Offers']] = avg_offers['Total']
df[['% Asian Offers', '% Black Offers', '% Latino Offers', '% White Offers']] = avg_offers[['Asian', 'Black', 'Hispanic', 'White']].div(avg_offers['Total'],axis='index').mul(100).round(1)
df[['% Asian Offers - Estimate Off By (+/-)', '% Black Offers - Estimate Off By (+/-)', '% Latino Offers - Estimate Off By (+/-)', '% White Offers - Estimate Off By (+/-)']] = max_offers.sub(avg_offers).div(avg_offers['Total'], axis='index').mul(100).round(1)[['Asian', 'Black','Hispanic','White']]

column_order = ['School Name']
column_order.extend(sorted([col for col in df if 'Applicants' in col]))
column_order.extend(sorted([col for col in df if 'Offers' in col]))

df[column_order].to_csv('../output/pct-applicants-offers-stats.csv')
df[column_order]

#### Combined: All Schools and Top Ranking Schools

df = pd.DataFrame()

top = metrics[metrics.eq('Far Above Average').sum(axis=1).gt(4)].index
top = [dbn for dbn in top if dbn in info.index]

groups = [('Top-Ranked Screened High Schools', top), 
          ('All Screened High Schools', info.index)]
rows = []

for (title, dbns) in groups:
    row = []
    a = avg_applicants.loc[dbns].sum()[['Total', 'Asian', 'Black', 'Hispanic', 'White']]
    o = avg_offers.loc[dbns].sum()[['Total', 'Asian', 'Black', 'Hispanic', 'White']]
    arr = max_applicants.loc[dbns].sum()[['Total', 'Asian', 'Black', 'Hispanic', 'White']]
    #print(arr - a)
    orr = max_offers.loc[dbns].sum()[['Total', 'Asian', 'Black', 'Hispanic', 'White']]
    print((arr - a).div(a['Total']).mul(100).round(1))
    print()
    print(count)
    
    row.append(title)
    row.append(len(dbns))
    row.append(a['Total'])
    row.extend(list(a[['Asian', 'Black', 'Hispanic', 'White']].div(a['Total']).mul(100).round(1)))
    row.extend(list((arr - a).div(a['Total']).mul(100).round(1)[['Asian', 'Black', 'Hispanic', 'White']]))
    row.append(o['Total'])
    row.extend(list(o[['Asian', 'Black', 'Hispanic', 'White']].div(o['Total']).mul(100).round(1)))
    row.extend(list((orr - o).div(o['Total']).mul(100).round(1)[['Asian', 'Black', 'Hispanic', 'White']]))

    rows.append(row)
    
columns = [
    'Category',
    '# Schools',
    '# Applicants',
    '% Asian Applicants', 
    '% Black Applicants', 
    '% Latino Applicants', 
    '% White Applicants',
    '% Asian Applicants - Estimate Off By (+/-)', 
    '% Black Applicants - Estimate Off By (+/-)', 
    '% Latino Applicants - Estimate Off By (+/-)', 
    '% White Applicants - Estimate Off By (+/-)',
    '# Offers',
    '% Asian Offers',
    '% Black Offers',
    '% Latino Offers',
    '% White Offers',
    '% Asian Offers - Estimate Off By (+/-)', 
    '% Black Offers - Estimate Off By (+/-)', 
    '% Latino Offers - Estimate Off By (+/-)', 
    '% White Offers - Estimate Off By (+/-)'
]

col_order = ['Category', '# Schools']
col_order.extend(sorted([col for col in columns if 'Applicants' in col]))
col_order.extend(sorted([col for col in columns if 'Offers' in col]))

df = pd.DataFrame(rows, columns=columns)[col_order]
df.to_csv('../output/all-vs-top-ranked-shares.csv', index=False)
df

pd.DataFrame(rows, columns=columns)[col_order].set_index('Category')

# "Top-Ranked" Screened Schools Demographics
d = pd.DataFrame()
a = avg_applicants.loc[top].sum()
a = a.rename({'Hispanic':'Latino'})
o = avg_offers.loc[top].sum()
o = o.rename({'Hispanic':'Latino'})
d['Applicants (%)'] = a.div(a['Total']).mul(100).round(1)[['Asian', 'Black', 'Latino', 'White']]
d['Offers (%)'] = o.div(o['Total']).mul(100).round(1)[['Asian', 'Black', 'Latino', 'White']]
d['Difference (%)'] = d['Offers (%)'].sub(d['Applicants (%)']).round(1)
d.to_csv('../output/top-ranked-demo-differences.csv')
display(d)
