==> notebooks/3-dia-statistics.ipynb <==
import pandas as pd
import numpy as np

info = pd.read_csv('../data/cleaned/info.csv')
directory = pd.read_csv('../data/2020_DOE_High_School_Directory.csv')
directory21 = pd.read_csv('../data/2021_DOE_High_School_Directory.csv')

applicants = pd.read_csv('../data/cleaned/applicants.csv')
applicants = applicants.set_index('dbn')
offers = pd.read_csv('../data/cleaned/offers.csv')
offers = offers.set_index('dbn')

dia = directory[['dbn', 'school_name', 'diversity_in_admissions']][
    directory['dbn'].isin(info['dbn']) & 
    ~(directory['diversity_in_admissions'].isna())]
dia = dia.set_index('dbn')

print("Of the {} screened schools in our dataset, {} participated in the Diversity in Admissions program.".format(len(info), len(dia)))

### Compare minimum thresholds to shares of applicants and offers.

dia_applicants = applicants.loc[dia.index]

dia_offers = offers.loc[dia.index]

dia['% Applicants'] = dia_applicants['Poverty'].div(dia_applicants['Total']).mul(100).round(1)
dia['% Offers'] = dia_offers['Poverty'].div(dia_offers['Total']).mul(100).round(1)
dia['% Change'] = dia['% Offers'] - dia['% Applicants']

dia['% Priority'] = dia['diversity_in_admissions'].apply(lambda s: s[s.find('%') - 2:s.find('%')])
dia['% Priority'] = pd.to_numeric(dia['% Priority'], errors='coerce')

dia

dia['diversity_in_admissions'].unique()

### Compare 2021 vs. 2020 DIA rates

dia21 = directory21[['dbn', 'school_name', 'diversity_in_admissions']][
    directory21['dbn'].isin(info['dbn']) & 
    ~(directory21['diversity_in_admissions'].isna())]
dia21 = dia21.set_index('dbn')

list(dia21.loc['02M416'])

dia

dia21['% Priority'] = dia21['diversity_in_admissions'].apply(lambda s: s[s.find('%') - 2:s.find('%')])
dia21['% Priority'] = pd.to_numeric(dia21['% Priority'], errors='coerce')

dia21_applicants = applicants.loc[dia21.index]
dia21_offers = offers.loc[dia21.index]
dia21['% Applicants'] = dia21_applicants['Poverty'].div(dia21_applicants['Total']).mul(100).round(1)
dia21['% Offers'] = dia21_offers['Poverty'].div(dia21_offers['Total']).mul(100).round(1)
dia21['% Change'] = dia21['% Offers'] - dia21['% Applicants']


list(dia21.loc['02M416'])

directory21['geoeligibility'].unique()

list(directory21.columns)

directory[[col for col in directory.columns if 'priorit' in col]].eq('Priority to continuing 8th graders').sum(axis=1).apply(bool).sum()

directory.loc[info.index][[col for col in directory.columns if 'priorit' in col]].eq('Priority to continuing 8th graders').sum(axis=1).apply(bool).sum()

len(directory)

directory21[[col for col in directory21.columns if 'priorit' in col]].eq('Priority to continuing 8th graders').sum(axis=1).apply(bool).sum()

len(directory21)

directory.loc[info.index][[col for col in directory.columns if 'priorit' in col]]

directory['admissionspriority11'].unique()
