==> notebooks/4-data-summary.ipynb <==
from collections import Counter

import pandas as pd
import numpy as np

metrics = pd.read_csv('../data/metrics_2019.csv')
metrics = metrics.set_index('dbn')

applicants = pd.read_csv('../data/cleaned/applicants.csv')
applicants = applicants.set_index('dbn')

offers = pd.read_csv('../data/cleaned/offers.csv')
offers.set_index('dbn')

info = pd.read_csv('../data/cleaned/info.csv')

metrics.loc[list(info['dbn'])].T.apply(Counter, axis=1).apply(pd.Series)[[
    'Far Below Average',
    'Below Average',
    'Average',
    'Above Average',
    'Far Above Average'
]].fillna(0).to_csv('../output/screened-school-ranking-breakdown.csv')

cols = ['Asian', 'Black', 'Latino', 'White', 'Other', 'Private']

a_redact = pd.DataFrame()
a_redact['Redactions for <6 Students'] = applicants.rename(columns={'Hispanic':'Latino'}).eq('s').sum()[cols]
a_redact['Redactions for 6+ Students'] = applicants.rename(columns={'Hispanic':'Latino'}).eq('s^').sum()[cols]
a_redact.to_csv('../output/redactions-applicants.csv')

o_redact = pd.DataFrame()
o_redact['Redactions for <6 Students'] = offers.rename(columns={'Hispanic':'Latino'}).eq('s').sum()[cols]
o_redact['Redactions for 6+ Students'] = offers.rename(columns={'Hispanic':'Latino'}).eq('s^').sum()[cols]
o_redact.to_csv('../output/redactions-offers.csv')
